Downstream projects can now require grunt and gulp tasks, no need to copy them anymore.
