--- package.gulp.json
@@ -1,45 +1,0 @@
-{
-  "name": "patternlab-node",
-  "description": "Pattern Lab is a collection of tools to help you create atomic design systems. This is the node command line interface (CLI).",
-  "version": "1.1.1",
-  "main": "./builder/patternlab.js",
-  "dependencies": {
-    "del": "^2.2.0",
-    "diveSync": "^0.3.0",
-    "fs-extra": "^0.26.5",
-    "glob": "^7.0.0",
-    "html-entities": "^1.2.0",
-    "mustache": "^2.2.1"
-  },
-  "devDependencies": {
-    "browser-sync": "^2.11.1",
-    "gulp": "^3.9.0",
-    "gulp-connect": "^2.3.1",
-    "gulp-copy": "0.0.2",
-    "gulp-header": "^1.7.1",
-    "gulp-load": "^0.1.1",
-    "gulp-nodeunit": "0.0.5",
-    "gulp-strip-banner": "0.0.2"
-  },
-  "keywords": [
-    "Pattern Lab",
-    "Atomic Web Design",
-    "Node",
-    "Grunt",
-    "Gulp",
-    "Javascript"
-  ],
-  "repository": {
-    "type": "git",
-    "url": "git://github.com/pattern-lab/patternlab-node.git"
-  },
-  "bugs": "https://github.com/pattern-lab/patternlab-node/issues",
-  "author": "Brian Muenzenmeyer",
-  "license": "MIT",
-  "scripts": {
-    "test": "grunt travis --verbose"
-  },
-  "engines": {
-    "node": ">=0.1"
-  }
-}
