+++ package.json
@@ -9,7 +9,8 @@
     "glob": "^7.0.0",
     "html-entities": "^1.2.0",
     "matchdep": "^1.0.0",
-    "mustache": "^2.2.0"
+    "mustache": "^2.2.1",
+    "del": "^2.2.0"
   },
   "devDependencies": {
     "bs-html-injector": "^3.0.0",
@@ -18,7 +19,15 @@
     "grunt-browser-sync": "^2.2.0",
     "grunt-contrib-copy": "^0.8.2",
     "grunt-contrib-nodeunit": "^0.4.1",
-    "grunt-contrib-watch": "^0.6.1"
+    "grunt-contrib-watch": "^0.6.1",
+    "browser-sync": "^2.11.1",
+    "gulp": "^3.9.0",
+    "gulp-connect": "^2.3.1",
+    "gulp-copy": "0.0.2",
+    "gulp-header": "^1.7.1",
+    "gulp-load": "^0.1.1",
+    "gulp-nodeunit": "0.0.5",
+    "gulp-strip-banner": "0.0.2"
   },
   "keywords": [
     "Pattern Lab",
